Refactor: further cleanup and small fixes
